+++ src/api/models/user-model.js
@@ -17,6 +17,15 @@ const findUserById = async (id) => {
   return rows[0];
 };
 
+const findUserByUsername = async (username) => {
+  const [rows] = await promisePool.execute(
+    'SELECT * FROM wsk_users WHERE username = ?',
+    [username]
+  );
+  if (rows.length === 0) return false;
+  return rows[0];
+};
+
 const addUser = async (user) => {
   const {name, username, email, password, role} = user;
   const sql = `INSERT INTO wsk_users (name, username, email, password, role)
@@ -30,38 +39,53 @@ const addUser = async (user) => {
   return {user_id: result[0].insertId};
 };
 
-const modifyUser = async (id, user) => {
-  const sql = promisePool.format(`UPDATE wsk_users SET ? WHERE user_id = ?`, [
-    user,
-    id,
-  ]);
-  const rows = await promisePool.execute(sql);
-  console.log('rows', rows);
-  if (rows[0].affectedRows === 0) {
+const modifyUser = async (id, user, authUser) => {
+  try {
+    let sql;
+    if (authUser.role && authUser.role === 'admin') {
+      sql = promisePool.format(`UPDATE wsk_users SET ? WHERE user_id = ?`, [user, id]);
+    } else if (authUser) {
+      sql = promisePool.format(`UPDATE wsk_users SET ? WHERE user_id = ? AND user_id = ?`, [user, id, authUser.user_id]);
+    } else {
+      return false;
+    }
+    const rows = await promisePool.execute(sql);
+    //console.log('rows', rows);
+    if (rows[0].affectedRows === 0) {
+      return false;
+    }
+    return {message: 'success'};
+  } catch (err) {
+    console.error(err);
     return false;
   }
-  return {message: 'sucess'}
 };
 
-const removeUser = async (id) => {
+const removeUser = async (id, authUser) => {
   try {
-    // first delete cat
-    await promisePool.execute(
-      'DELETE FROM wsk_cats WHERE owner = ?', [id]
-    );
-    const [rows] = await promisePool.execute(
-      'DELETE FROM wsk_users WHERE user_id = ?', [id]
-    );
+    if (!(authUser.role && authUser.role === 'admin') && String(authUser.user_id) !== String(id)) {
+      return false;
+    }
+ 
+    await promisePool.execute('DELETE FROM wsk_cats WHERE owner = ?', [id]);
+
+    let result;
+    if (authUser && authUser.role && authUser.role === 'admin') {
+      result = await promisePool.execute('DELETE FROM wsk_users WHERE user_id = ?', [id]);
+    } else {
+      result = await promisePool.execute('DELETE FROM wsk_users WHERE user_id = ? AND user_id = ?', [id, authUser.user_id]);
+    }
+    const rows = result[0];
     console.log('rows', rows);
     if (rows.affectedRows === 0) {
       return false;
     }
     return {message: 'success'};
   } catch (error) {
-    console.error('Error removing user:', error);
+    console.error('Error;', error);
     return false;
   }
 };
 
 
-export {listAllUsers, findUserById, addUser, modifyUser, removeUser};
+export {listAllUsers, findUserById, addUser, modifyUser, removeUser, findUserByUsername};
